feat: refactor visual block handling by moving utility functions to visual_magic_model_utils.py
